refactor: split sqlite database service
